// product-plan/sections/homepage/types.ts
export type PersonaType = 'member' | 'organizer' | 'both'

export interface Benefit {
  id: string
  title: string
  description: string
  icon: string
  forPersona: PersonaType
}